[ADD] DTO & global pipes

+++ src/controller/PostController.ts
@@ -4,9 +4,8 @@ import { PostBoard } from "entity/post";
 import { PostService } from "service/PostService";
 import { disposeError } from 'lib/DisposeError';
 import { handleSuccess } from 'lib/Response/handleSuccess';
-import { IPostTypes } from 'types/post';
-import { validateCreatePost } from 'validation/Post';
-import { handleFailed } from 'lib/Response/handleFailed';
+import { PostDto } from 'entity/dto/Post.dto';
+import HttpError from 'exception/HttpError';
 
 @Controller('post')
 export class PostController {
@@ -27,6 +26,10 @@ export class PostController {
   @Get('/:idx')
   public async getPost(@Param('idx') idx: number, @Res() response: Response) {
     try {
+      if (!Number.isInteger(idx)) {
+        throw new HttpError(400, '검증 오류입니다.');
+      }
+      
       const post: PostBoard = await this.postService.getPost(idx);
       handleSuccess(response, 200, '글을 조회하였습니다.', { post });
       return;
@@ -36,12 +39,8 @@ export class PostController {
   }
 
   @Post('/')
-  public async createPost(@Body() request: IPostTypes, @Res() response: Response) {
+  public async createPost(@Body() request: PostDto, @Res() response: Response) {
     try {
-      if (!validateCreatePost(request, response)) {
-        return;
-      }
-
       await this.postService.createPost(request);
       handleSuccess(response, 200, '글을 작성하였습니다.');
       return;
@@ -56,15 +55,6 @@ export class PostController {
     try {
       const postIdx: number = Number(request.params.idx);
 
-      if (!Number.isInteger(postIdx)) {
-        handleFailed(response, 400, '검증 오류입니다.');
-        return;
-      }
-
-      if (!validateCreatePost(request, response)) {
-        return;
-      }
-
       await this.postService.modifyPost(postIdx, request.body);
       handleSuccess(response, 200, '글을 수정하였습니다.');
       return;
@@ -76,6 +66,10 @@ export class PostController {
   @Delete('/:idx')
   public async deletePost(@Param('idx') idx: number, @Res() response: Response) {
     try {
+      if (!Number.isInteger(idx)) {
+        throw new HttpError(400, '검증 오류입니다.');
+      }
+
       await this.postService.deletePost(idx);
       handleSuccess(response, 200, '글을 삭제하였습니다.');
       return;
